pyproject.toml
[tool.poetry]
name = "arxprism"
version = "0.1.0"
description = "Academic knowledge graph extraction pipeline from arXiv papers"
authors = ["ArxPrism Team"]
readme = "README.md"
packages = [{include = "src", from = "."}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.30.0"}
python-multipart = "^0.0.9"
pydantic = "^2.9.0"
pydantic-settings = "^2.6.0"
neo4j = "^5.25.0"
openai = "^1.55.0"
celery = "^5.4.0"
redis = "^5.0.0"
arxiv = "^2.1.0"
beautifulsoup4 = "^4.12.0"
markdownify = "^0.14.1"
python-dotenv = "^1.0.0"
httpx = "^0.27.0"
supabase = "^2.6.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
streamlit = "^1.40.0"
streamlit-agraph = "^0.0.45"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.0"
pytest-asyncio = "^0.24.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
